Make group number limit for khyperloglog_agg feature config controlled
